Move "Two" controllers & transformers alongside everything else.
