+++ routes/api.php
@@ -18,11 +18,11 @@ $router->group(['prefix' => 'v2', 'as' => 'v2.'], function () {
     $this->get('userinfo', 'UserInfoController@show');
 
     // Users
-    $this->resource('users', 'Two\UserController');
-    // Get user by mobile #
-    $this->get('mobile/{mobile}', 'Two\MobileController@show');
-    // Get user by email
-    $this->get('email/{email}', 'Two\EmailController@show');
+    $this->resource('users', 'UserController');
+
+    // User (by email or mobile number)
+    $this->get('mobile/{mobile}', 'MobileController@show');
+    $this->get('email/{email}', 'EmailController@show');
 
     // Profile
     // ...
